feat: CommonLaundry클래스 weight 타입 Integer->Double수정

+++ src/main/java/aug/laundry/domain/CommonLaundry.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 public class CommonLaundry {
 
     private Long commonLaundryId;
-    private Integer commonLaundryWeight;
+    private Double commonLaundryWeight;
     private Long ordersDetailId;
 
     public CommonLaundry() {
